dom: add focus/selection tracker pattern
The broad goal is to be able to build horizontal radio buttons in a group with custom selection rendering.

The approach is to build a "tracker" state which holds the key of *current* active item. The main work in dom/tracker is building sub-streams for each individual item.  The substreams are all boolean streams where setting one to true will cause the parent tracker stream to have its selection updated accordingly.

The code in todo/controls/filter.go has been updated to use the new approach to render the todo filters into ("All", "Active", "Done") categories which leads to a prettier UX.

The stream/sub-stream pattern seems quite powerful but it is rather unwieldy to build the datastructures needed for it because the primitives are not quite clear yet.  It would not make sense to hand-roll all the specific stream implementations written so far but I think with a good set of primitives all of these can be code-generated instead.